xl: heal bucket should validate if bucket exists first.
Fixes #2944
